flask_ml_api/requirements.txt
flask
flask-cors
opencv-contrib-python-headless
numpy
scikit-learn
scikit-image
joblib
Pillow
gunicorn
